+++ examples/data-sources/fmc_application_categories/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_application_categories" "example" {
+  items = {
+    "software update" = {
+    }
+  }
+}