+++ examples/data-sources/fmc_application_types/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_application_types" "example" {
+  items = {
+    "Webapp" = {
+    }
+  }
+}