+++ examples/data-sources/fmc_applications/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_applications" "example" {
+  items = {
+    "10Web" = {
+    }
+  }
+}