+++ internal/provider/provider.go
@@ -293,6 +293,7 @@ func (p *FmcProvider) Configure(ctx context.Context, req provider.ConfigureReque
 func (p *FmcProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewAccessControlPolicyResource,
+		NewApplicationFilterResource,
 		NewBFDTemplateResource,
 		NewDeviceResource,
 		NewDeviceBFDResource,
@@ -362,6 +363,19 @@ func (p *FmcProvider) Resources(ctx context.Context) []func() resource.Resource 
 func (p *FmcProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewAccessControlPolicyDataSource,
+		NewApplicationDataSource,
+		NewApplicationBusinessRelevanceDataSource,
+		NewApplicationBusinessRelevancesDataSource,
+		NewApplicationCategoriesDataSource,
+		NewApplicationCategoryDataSource,
+		NewApplicationFilterDataSource,
+		NewApplicationRiskDataSource,
+		NewApplicationRisksDataSource,
+		NewApplicationTagDataSource,
+		NewApplicationTagsDataSource,
+		NewApplicationTypeDataSource,
+		NewApplicationTypesDataSource,
+		NewApplicationsDataSource,
 		NewBFDTemplateDataSource,
 		NewDeviceDataSource,
 		NewDeviceBFDDataSource,